restructure and refactor some code to be more understandable. mainly naming
