+++ lib/common/widgets/flower/flower_menu.dart
@@ -3,8 +3,8 @@ import 'package:redux/redux.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower/icon_circle.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower/flower.dart';
+import 'package:learning_compass_exp/common/widgets/flower/icon_circle.dart';
+import 'package:learning_compass_exp/common/widgets/flower/flower.dart';
 import 'package:learning_compass_exp/store/actions/actions.dart';
