+++ lib/routes.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:learning_compass_exp/screens/about/about.dart';
 import 'package:learning_compass_exp/screens/home/index.dart';
-import 'package:learning_compass_exp/screens/info/index.dart';
+import 'package:learning_compass_exp/screens/construct_info/chapter_index.dart';
 import 'package:learning_compass_exp/data/constants/PETAL_CONSTANTS.dart';
