+++ lib/screens/construct_info/widgets/chapter_screen/chapter_screen_body.dart
@@ -6,8 +6,8 @@ import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/models/chapter_state.dart';
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/store/actions/actions.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_content.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/magic_word_bar.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/chapter_content.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/magic_word_bar.dart';
 
 class ChapterScreenBody extends StatelessWidget {
   final Map<String, dynamic> chapter;
