+++ lib/screens/construct_info/widgets/chapter_screen/star_info_dialog.dart
