+++ lib/screens/home/index.dart
@@ -3,7 +3,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:learning_compass_exp/screens/home/home.dart';
 import 'package:learning_compass_exp/screens/home/widgets/drawer/drawer.dart';
-import 'package:learning_compass_exp/screens/home/welcome_dialog.dart';
+import 'package:learning_compass_exp/screens/home/widgets/welcome_dialog.dart';
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/store/actions/actions.dart';
