+++ test/widget_tests/home/icon_circle_widget_test.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/app.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower/petal_icon_button.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower/icon_circle.dart';
+import 'package:learning_compass_exp/common/widgets/flower/petal_icon_button.dart';
+import 'package:learning_compass_exp/common/widgets/flower/icon_circle.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 import 'package:learning_compass_exp/common/widgets/custom_icons_icons.dart';
