+++ test/widget_tests/home/progress_bar_widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:learning_compass_exp/screens/home/widgets/flower/icon_popup/progress_bar.dart';
+import 'package:learning_compass_exp/common/widgets/flower/icon_popup/progress_bar.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/constants/PETAL_CONSTANTS.dart';
