+++ test/widget_tests/info/chapter_content_test.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/gestures.dart';
 
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_content.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/word_found_dialog.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/chapter_content.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/models/petal_names.dart';
 import 'package:learning_compass_exp/data/models/chapter_state.dart';
