+++ test/widget_tests/info/chapter_index_test.dart
@@ -4,8 +4,8 @@ import 'package:redux/redux.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
-import 'package:learning_compass_exp/screens/info/chapter_index/chapter_index.dart';
-import 'package:learning_compass_exp/screens/info/chapter_index/chapter_index_card.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_index/chapter_index.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_index/chapter_index_card.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/models/petal_names.dart';
