+++ test/widget_tests/info/chapter_screen_body_test.dart
@@ -5,8 +5,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'package:mockito/mockito.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_screen_body.dart';
-import 'package:learning_compass_exp/screens/info/chapter_index/chapter_index.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/chapter_screen_body.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/models/petal_names.dart';
