+++ test/widget_tests/info/chapter_screen_test.dart
@@ -3,9 +3,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:redux/redux.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_screen.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_selection_bar.dart';
-import 'package:learning_compass_exp/screens/info/chapter_screen/chapter_screen_body.dart';
+import 'package:learning_compass_exp/screens/construct_info/chapter_screen.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/chapter_selection_bar.dart';
+import 'package:learning_compass_exp/screens/construct_info/widgets/chapter_screen/chapter_screen_body.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 import 'package:learning_compass_exp/data/models/petal_names.dart';
 import 'package:learning_compass_exp/store/app_state.dart';
